chore(core): move tests to core
these were still in instantsearch.js but actually are for the "core" domain now.
