+++ packages/instantsearch-core/src/__tests__/RoutingManager.test.ts
@@ -6,12 +6,15 @@ import { createSearchClient } from '@instantsearch/mocks';
 import { wait } from '@instantsearch/testutils/wait';
 import qs from 'qs';
 
-import instantsearch from '../..';
-import { createWidget } from '../../../test/createWidget';
-import { connectHitsPerPage, connectSearchBox } from '../../connectors';
-import historyRouter from '../routers/history';
+import {
+  instantsearch,
+  historyRouter,
+  connectHitsPerPage,
+  connectSearchBox,
+} from '..';
+import { createWidget } from '../../test/createWidget';
 
-import type { Router, UiState, StateMapping, IndexUiState } from '../../types';
+import type { Router, UiState, StateMapping, IndexUiState } from '../types';
 import type { JSDOM } from 'jsdom';
 
 declare const jsdom: JSDOM;
