+++ packages/instantsearch-core/src/routing/__tests__/dispose-start.test.ts
@@ -5,21 +5,29 @@
 import { createSearchClient } from '@instantsearch/mocks';
 import { wait } from '@instantsearch/testutils/wait';
 
-import instantsearch from '../../..';
-import { connectSearchBox } from '../../../connectors';
-import historyRouter from '../../routers/history';
+import { instantsearch, historyRouter, connectSearchBox } from '../..';
+
+import type { InstantSearch } from '../..';
 
 /* eslint no-lone-blocks: "off" */
 
 const writeDelay = 10;
 const writeWait = 10 * writeDelay;
 
-describe('routing with no navigation', () => {
-  test('cleans the URL when InstantSearch is disposed within the same page', async () => {
+const addWidgetsAndStart = (search: InstantSearch) => {
+  search.addWidgets([connectSearchBox(() => {})({})]);
+  search.start();
+};
+
+describe('routing back and forth to an InstantSearch instance', () => {
+  test('updates the URL after the instance is disposed then restarted', async () => {
     // -- Flow
     // 1. Initial: '/'
     // 2. Refine: '/?indexName[query]=Apple'
     // 3. Dispose: '/'
+    // 4. Refine: '/'
+    // 5. Start: '/'
+    // 6. Refine: '/?indexName[query]=Apple'
 
     const pushState = jest.spyOn(window.history, 'pushState');
 
@@ -35,8 +43,7 @@ describe('routing with no navigation', () => {
 
     // 1. Initial: '/'
     {
-      search.addWidgets([connectSearchBox(() => {})({})]);
-      search.start();
+      addWidgetsAndStart(search);
 
       await wait(writeWait);
       expect(window.location.search).toEqual('');
@@ -48,7 +55,6 @@ describe('routing with no navigation', () => {
       search.renderState.indexName.searchBox!.refine('Apple');
 
       await wait(writeWait);
-      expect(window.location.pathname).toEqual('/');
       expect(window.location.search).toEqual(
         `?${encodeURI('indexName[query]=Apple')}`
       );
@@ -60,9 +66,37 @@ describe('routing with no navigation', () => {
       search.dispose();
 
       await wait(writeWait);
-      expect(window.location.pathname).toEqual('/');
       expect(window.location.search).toEqual('');
       expect(pushState).toHaveBeenCalledTimes(2);
+    }
+
+    // 4. Refine: '/'
+    {
+      search.renderState.indexName.searchBox!.refine('Apple');
+
+      await wait(writeWait);
+      expect(window.location.search).toEqual('');
+      expect(pushState).toHaveBeenCalledTimes(2);
+    }
+
+    // 5. Start: '/'
+    {
+      addWidgetsAndStart(search);
+
+      await wait(writeWait);
+      expect(window.location.search).toEqual('');
+      expect(pushState).toHaveBeenCalledTimes(2);
+    }
+
+    // 6. Refine: '/?indexName[query]=Apple'
+    {
+      search.renderState.indexName.searchBox!.refine('Samsung');
+
+      await wait(writeWait);
+      expect(window.location.search).toEqual(
+        `?${encodeURI('indexName[query]=Samsung')}`
+      );
+      expect(pushState).toHaveBeenCalledTimes(3);
     }
   });
 });
