+++ packages/instantsearch-core/src/routing/__tests__/external-influence.test.ts
@@ -5,31 +5,20 @@
 import { createSearchClient } from '@instantsearch/mocks';
 import { wait } from '@instantsearch/testutils/wait';
 
-import instantsearch from '../../..';
-import { connectSearchBox } from '../../../connectors';
-import historyRouter from '../../routers/history';
-
-import type InstantSearch from '../../InstantSearch';
+import { instantsearch, historyRouter, connectSearchBox } from '../..';
 
 /* eslint no-lone-blocks: "off" */
 
 const writeDelay = 10;
 const writeWait = 10 * writeDelay;
 
-const addWidgetsAndStart = (search: InstantSearch) => {
-  search.addWidgets([connectSearchBox(() => {})({})]);
-  search.start();
-};
-
-describe('routing back and forth to an InstantSearch instance', () => {
-  test('updates the URL after the instance is disposed then restarted', async () => {
+describe('routing with external influence', () => {
+  test('keeps on working when the URL is updated by another program', async () => {
     // -- Flow
     // 1. Initial: '/'
     // 2. Refine: '/?indexName[query]=Apple'
-    // 3. Dispose: '/'
-    // 4. Refine: '/'
-    // 5. Start: '/'
-    // 6. Refine: '/?indexName[query]=Apple'
+    // 3. External influence: '/about'
+    // 4. Refine: '/about?indexName[query]=Samsung'
 
     const pushState = jest.spyOn(window.history, 'pushState');
 
@@ -45,7 +34,8 @@ describe('routing back and forth to an InstantSearch instance', () => {
 
     // 1. Initial: '/'
     {
-      addWidgetsAndStart(search);
+      search.addWidgets([connectSearchBox(() => {})({})]);
+      search.start();
 
       await wait(writeWait);
       expect(window.location.search).toEqual('');
@@ -63,42 +53,25 @@ describe('routing back and forth to an InstantSearch instance', () => {
       expect(pushState).toHaveBeenCalledTimes(1);
     }
 
-    // 3. Dispose: '/'
+    // 3. External influence: '/about'
     {
-      search.dispose();
+      window.history.pushState({}, '', '/about');
 
       await wait(writeWait);
+      expect(window.location.pathname).toEqual('/about');
       expect(window.location.search).toEqual('');
       expect(pushState).toHaveBeenCalledTimes(2);
     }
 
-    // 4. Refine: '/'
-    {
-      search.renderState.indexName.searchBox!.refine('Apple');
-
-      await wait(writeWait);
-      expect(window.location.search).toEqual('');
-      expect(pushState).toHaveBeenCalledTimes(2);
-    }
-
-    // 5. Start: '/'
-    {
-      addWidgetsAndStart(search);
-
-      await wait(writeWait);
-      expect(window.location.search).toEqual('');
-      expect(pushState).toHaveBeenCalledTimes(2);
-    }
-
-    // 6. Refine: '/?indexName[query]=Apple'
+    // 4. Refine: '/about?indexName[query]=Samsung'
     {
       search.renderState.indexName.searchBox!.refine('Samsung');
 
       await wait(writeWait);
+      expect(window.location.pathname).toEqual('/about');
       expect(window.location.search).toEqual(
         `?${encodeURI('indexName[query]=Samsung')}`
       );
-      expect(pushState).toHaveBeenCalledTimes(3);
     }
   });
 });
