+++ packages/instantsearch-core/src/routing/__tests__/spa-debounced.test.ts
@@ -5,22 +5,22 @@
 import { createSearchClient } from '@instantsearch/mocks';
 import { wait } from '@instantsearch/testutils/wait';
 
-import instantsearch from '../../..';
-import { connectSearchBox } from '../../../connectors';
-import historyRouter from '../../routers/history';
+import { instantsearch, historyRouter, connectSearchBox } from '../..';
 
 /* eslint no-lone-blocks: "off" */
 
 const writeDelay = 10;
 const writeWait = 10 * writeDelay;
 
-describe('routing with external influence', () => {
-  test('keeps on working when the URL is updated by another program', async () => {
+describe('routing with debounced third-party client-side router', () => {
+  test('does not clean the URL after navigating', async () => {
     // -- Flow
     // 1. Initial: '/'
     // 2. Refine: '/?indexName[query]=Apple'
-    // 3. External influence: '/about'
-    // 4. Refine: '/about?indexName[query]=Samsung'
+    // 3. Dispose: '/'
+    // 4. Route change: '/about'
+    // 5. Back: '/'
+    // 6. Back: '/?indexName[query]=Apple'
 
     const pushState = jest.spyOn(window.history, 'pushState');
 
@@ -55,25 +55,45 @@ describe('routing with external influence', () => {
       expect(pushState).toHaveBeenCalledTimes(1);
     }
 
-    // 3. External influence: '/about'
+    // 3. Dispose: '/'
+    {
+      search.dispose();
+
+      await wait(writeWait);
+      expect(window.location.pathname).toEqual('/');
+      expect(window.location.search).toEqual('');
+      expect(pushState).toHaveBeenCalledTimes(2);
+    }
+
+    // 4. Route change: '/about'
     {
       window.history.pushState({}, '', '/about');
 
       await wait(writeWait);
       expect(window.location.pathname).toEqual('/about');
       expect(window.location.search).toEqual('');
-      expect(pushState).toHaveBeenCalledTimes(2);
+      expect(pushState).toHaveBeenCalledTimes(3);
     }
 
-    // 4. Refine: '/about?indexName[query]=Samsung'
+    // 5. Back: '/'
     {
-      search.renderState.indexName.searchBox!.refine('Samsung');
+      window.history.back();
 
       await wait(writeWait);
-      expect(window.location.pathname).toEqual('/about');
+      expect(window.location.pathname).toEqual('/');
+      expect(window.location.search).toEqual('');
+    }
+
+    // 6. Back: '/?indexName[query]=Apple'
+    {
+      window.history.back();
+
+      await wait(writeWait);
+      expect(window.location.pathname).toEqual('/');
       expect(window.location.search).toEqual(
-        `?${encodeURI('indexName[query]=Samsung')}`
+        `?${encodeURI('indexName[query]=Apple')}`
       );
+      expect(pushState).toHaveBeenCalledTimes(3);
     }
   });
 });
