+++ packages/instantsearch-core/src/routing/__tests__/spa.test.ts
@@ -5,24 +5,22 @@
 import { createSearchClient } from '@instantsearch/mocks';
 import { wait } from '@instantsearch/testutils/wait';
 
-import instantsearch from '../../..';
-import { connectSearchBox } from '../../../connectors';
-import historyRouter from '../../routers/history';
+import { instantsearch, historyRouter, connectSearchBox } from '../..';
 
 /* eslint no-lone-blocks: "off" */
 
 const writeDelay = 10;
 const writeWait = 10 * writeDelay;
 
-describe('routing with debounced third-party client-side router', () => {
+describe('routing with third-party client-side router', () => {
   test('does not clean the URL after navigating', async () => {
     // -- Flow
     // 1. Initial: '/'
     // 2. Refine: '/?indexName[query]=Apple'
-    // 3. Dispose: '/'
-    // 4. Route change: '/about'
-    // 5. Back: '/'
-    // 6. Back: '/?indexName[query]=Apple'
+    // 3. Navigate: '/about'
+    // 4. Back: '/?indexName[query]=Apple'
+    // 5. Restart: '/?indexName[query]=Apple'
+    // 6. Refine: '/?indexName[query]=Samsung'
 
     const pushState = jest.spyOn(window.history, 'pushState');
 
@@ -57,36 +55,18 @@ describe('routing with debounced third-party client-side router', () => {
       expect(pushState).toHaveBeenCalledTimes(1);
     }
 
-    // 3. Dispose: '/'
+    // 3. Navigate: '/about'
     {
       search.dispose();
-
-      await wait(writeWait);
-      expect(window.location.pathname).toEqual('/');
-      expect(window.location.search).toEqual('');
-      expect(pushState).toHaveBeenCalledTimes(2);
-    }
-
-    // 4. Route change: '/about'
-    {
       window.history.pushState({}, '', '/about');
 
       await wait(writeWait);
       expect(window.location.pathname).toEqual('/about');
       expect(window.location.search).toEqual('');
-      expect(pushState).toHaveBeenCalledTimes(3);
+      expect(pushState).toHaveBeenCalledTimes(2);
     }
 
-    // 5. Back: '/'
-    {
-      window.history.back();
-
-      await wait(writeWait);
-      expect(window.location.pathname).toEqual('/');
-      expect(window.location.search).toEqual('');
-    }
-
-    // 6. Back: '/?indexName[query]=Apple'
+    // 4. Back to previous page
     {
       window.history.back();
 
@@ -94,6 +74,28 @@ describe('routing with debounced third-party client-side router', () => {
       expect(window.location.pathname).toEqual('/');
       expect(window.location.search).toEqual(
         `?${encodeURI('indexName[query]=Apple')}`
+      );
+    }
+
+    // 5. Restart InstantSearch
+    {
+      search.addWidgets([connectSearchBox(() => {})({})]);
+      search.start();
+
+      await wait(writeWait);
+      expect(window.location.pathname).toEqual('/');
+      expect(window.location.search).toEqual(
+        `?${encodeURI('indexName[query]=Apple')}`
+      );
+    }
+
+    // 6. Refine: '/?indexName[query]=Samsung'
+    {
+      search.renderState.indexName.searchBox!.refine('Samsung');
+
+      await wait(writeWait);
+      expect(window.location.search).toEqual(
+        `?${encodeURI('indexName[query]=Samsung')}`
       );
       expect(pushState).toHaveBeenCalledTimes(3);
     }
